tests/requirements.txt
# 테스트 실행에 필요한 외부 패키지 의존성
# pip install -r requirements.txt로 설치

# Redis 클라이언트 라이브러리
redis>=4.5.1

# RabbitMQ 클라이언트 라이브러리
pika>=1.3.1

# 테스트 프레임워크
pytest>=7.3.1

# HTTP 요청 라이브러리 
requests>=2.28.2

# 환경 변수 로딩 라이브러리
python-dotenv>=1.0.0 